Added "I'm Feeling Lucky" button
Still needed to add the actual code for the button.
Also, had to change database version as Android complained.

+++ app/src/main/java/com/example/encounter_generator/database/XPbyCREntity.kt
@@ -6,4 +6,4 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "challenge_ratings")
 data class XPbyCR(@PrimaryKey @ColumnInfo(name = "id") val id : Int, @ColumnInfo(name = "challenge") val CR: String,
-                  @ColumnInfo(name = "XP") val XP : Double)
+                  @ColumnInfo(name = "XP") val XP : Int)
